Added keybind to allow zeus to teleport on L Ctrl + H
Adapted from another script it have, Been using it for a while, Even tho i can use the context menu or a module I find using it as a keybind is just by far easier, Thought might as well add to base ezm.
Im not really sure there is anything i could change to the actual script as it is pretty basic so i left it as is and just adapted it to the differnt functions.